Van parts package renaming
