+++ robocode.api/src/main/java/foobar/hippy/van/booster/SuperBooster.java
@@ -1,4 +1,4 @@
-package foobar.booster;
+package foobar.hippy.van.booster;
 
 /**
  * SuperBooster is a booster.
