+++ robocode.api/src/main/java/foobar/hippy/van/booster/UltraBooster.java
@@ -1,4 +1,4 @@
-package foobar.booster;
+package foobar.hippy.van.booster;
 
 /**
  * UltraBooster is a booster.
