+++ robocode.api/src/main/java/foobar/hippy/van/coolingsystem/BigFan.java
@@ -1,4 +1,4 @@
-package foobar.coolingsystem;
+package foobar.hippy.van.coolingsystem;
 
 /**
  * An BigFan cooling system.
