+++ robocode.api/src/main/java/foobar/hippy/van/coolingsystem/Fan.java
@@ -1,4 +1,4 @@
-package foobar.coolingsystem;
+package foobar.hippy.van.coolingsystem;
 
 /**
  * An Fan cooling system.
