+++ robocode.api/src/main/java/foobar/hippy/van/coolingsystem/LiquidHydrogen.java
@@ -1,4 +1,4 @@
-package foobar.coolingsystem;
+package foobar.hippy.van.coolingsystem;
 
 /**
  * A LiquidHydrogen cooling system.
