+++ robocode.api/src/main/java/foobar/hippy/van/fuel/Coal.java
@@ -1,4 +1,4 @@
-package foobar.fuel;
+package foobar.hippy.van.fuel;
 
 /**
  * Coal is a fuel.
