+++ robocode.api/src/main/java/foobar/hippy/van/fuel/FuelDecorator.java
@@ -1,4 +1,4 @@
-package foobar.fuel;
+package foobar.hippy.van.fuel;
 
 /**
  * FuelDecorator acts as abstract decorator in decorator pattern.
