+++ robocode.api/src/main/java/foobar/hippy/van/fuel/IFuel.java
@@ -1,4 +1,4 @@
-package foobar.fuel;
+package foobar.hippy.van.fuel;
 
 /**
  * Fuel interface.
