+++ robocode.samples/src/main/java/foobar/race/samples/Technora1b.java
@@ -1,9 +1,9 @@
 package foobar.race.samples;
 
-import foobar.fuel.Coal;
-import foobar.fuel.Hydrogen;
-import foobar.fuel.Petrol;
 import foobar.hippy.van.AbstractVan;
+import foobar.hippy.van.fuel.Coal;
+import foobar.hippy.van.fuel.Hydrogen;
+import foobar.hippy.van.fuel.Petrol;
 import robocode.ScannedRobotEvent;
 
 /**
