+++ robocode.samples/src/main/java/foobar/race/samples/Technora1d.java
@@ -2,10 +2,10 @@ package foobar.race.samples;
 
 import java.awt.Color;
 
-import foobar.fuel.Coal;
-import foobar.fuel.Hydrogen;
-import foobar.fuel.Petrol;
 import foobar.hippy.van.AbstractVan;
+import foobar.hippy.van.fuel.Coal;
+import foobar.hippy.van.fuel.Hydrogen;
+import foobar.hippy.van.fuel.Petrol;
 import robocode.ScannedRobotEvent;
 
 /**
