+++ robocode.samples/src/main/java/foobar/race/samples/Technora2a.java
@@ -1,10 +1,10 @@
 package foobar.race.samples;
 
-import foobar.coolingsystem.Fan;
-import foobar.fuel.Coal;
-import foobar.fuel.Hydrogen;
-import foobar.fuel.Petrol;
 import foobar.hippy.van.AbstractVan;
+import foobar.hippy.van.coolingsystem.Fan;
+import foobar.hippy.van.fuel.Coal;
+import foobar.hippy.van.fuel.Hydrogen;
+import foobar.hippy.van.fuel.Petrol;
 import robocode.ScannedRobotEvent;
 
 /**
