+++ robocode.samples/src/main/java/foobar/race/samples/Technora2d.java
@@ -1,11 +1,11 @@
 package foobar.race.samples;
 
-import foobar.coolingsystem.Fan;
-import foobar.fuel.Coal;
-import foobar.fuel.Hydrogen;
-import foobar.fuel.Petrol;
 import foobar.hippy.van.AbstractVan;
-import foobar.model.Engine;
+import foobar.hippy.van.coolingsystem.Fan;
+import foobar.hippy.van.engine.Engine;
+import foobar.hippy.van.fuel.Coal;
+import foobar.hippy.van.fuel.Hydrogen;
+import foobar.hippy.van.fuel.Petrol;
 import robocode.ScannedRobotEvent;
 
 /**
