+++ robocode.samples/src/main/java/foobar/race/samples/Technora3a.java
@@ -2,11 +2,11 @@ package foobar.race.samples;
 
 import java.awt.Color;
 
-import foobar.coolingsystem.LiquidHydrogen;
-import foobar.fuel.Coal;
-import foobar.fuel.Hydrogen;
-import foobar.fuel.Petrol;
 import foobar.hippy.van.AbstractVan;
+import foobar.hippy.van.coolingsystem.LiquidHydrogen;
+import foobar.hippy.van.fuel.Coal;
+import foobar.hippy.van.fuel.Hydrogen;
+import foobar.hippy.van.fuel.Petrol;
 import robocode.ScannedRobotEvent;
 
 /**
